Backend/requirements.txt
asgiref>=3.11.0
certifi>=2025.11.12
charset-normalizer>=3.4.4
Django>=6.0
django-cors-headers>=4.9.0
djangorestframework>=3.16.1
djangorestframework_simplejwt>=5.5.1
idna>=3.11
psycopg2-binary>=2.9.11
PyJWT>=2.10.1
python-dotenv>=1.2.1
requests>=2.32.5
sqlparse>=0.5.4
urllib3>=2.6.2
mercadopago>=2.2.0
pydantic>=2.7.2
tenacity>=8.2.3
drf-spectacular>=0.27.0
gunicorn>=21.2.0
Pillow>=10.2.0